Configure mintmaker to avoid bumping ose-operator-registry
The tag specified for the ose-operator-registry image,
is also implicitly used to select the target OCP
minor we intend to release to.
See: https://konflux-ci.dev/architecture/ADR/0026-specifying-ocp-targets-for-fbc.html
Let's ensure that mintmaker is not going to bump
it to ensure we are going to release on the
expected OCP minor.

Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>